reformated descriptorSets and add .idea to git ignore
