+++ .gitignore
@@ -71,3 +71,4 @@ logs/
 users/
 tomcat_logs/
 
+.idea
