--- .idea/artifacts/chembench_war.xml
@@ -1,8 +1,0 @@
-<component name="ArtifactManager">
-  <artifact type="war" name="chembench:war">
-    <output-path>$PROJECT_DIR$/target</output-path>
-    <root id="archive" name="chembench-2.2-SNAPSHOT.war">
-      <element id="artifact" artifact-name="chembench:war exploded" />
-    </root>
-  </artifact>
-</component>
