+++ src/main/java/edu/unc/ceccr/chembench/global/Constants.java
@@ -15,6 +15,9 @@ public class Constants {
         AUTOSCALING, RANGESCALING, NOSCALING
     }
 
+    public static final String[] DESCRIPTORSETS = new String[]{Constants.CDK, Constants.DRAGONH,
+            Constants.DRAGONNOH, Constants.MOE2D, Constants.MACCS, Constants.ISIDA, Constants.DRAGON7 };
+
     //After submitting around 250 jobs, LSF won't let you submit more, it just returns an error.
     //Cap the number of jobs that can be added to the LSF queue at any given time.
     public static final int MAXLSFJOBS = 200;
@@ -210,16 +213,4 @@ public class Constants {
     // modelability index threshold
     public static double MODI_MODELABLE = 0.65;
 
-    //dictionary structure for mapping descriptor constants
-    //to their file endings at position 0
-    //to their error file at position 1
-    public static final Map<String, String[]> descriptorToFileEnding = ImmutableMap.<String, String[]>builder()
-            .put(Constants.CDK, new String[]{".cdk", "cdk.out"})
-            .put(Constants.DRAGONH, new String[]{".dragonH", "dragonH.out"})
-            .put(Constants.DRAGONNOH, new String[]{".dragonNoH", "dragonNoH.out"})
-            .put(Constants.MOE2D, new String[]{".moe2D", "moe2d.out"})
-            .put(Constants.MACCS, new String[]{".maccs" , "maccs.out"})
-            .put(Constants.ISIDA, new String[]{".ISIDA", "ISIDA.out"})
-            .put(Constants.DRAGON7, new String[]{".dragon7", "dragon7.out"})
-            .build();
 }
